requirements.txt
dnspython==2.2.0
pythonwhois==2.4.3
